+++ src/main/java/tornadofx/Collections.kt
@@ -149,7 +149,7 @@ fun <T> MutableList<T>.swap(indexOne: Int, indexTwo: Int) {
  */
 fun <T> MutableList<T>.swap(itemOne: T, itemTwo: T) = swap(indexOf(itemOne),indexOf(itemTwo))
 
-fun <SourceType, TargetType> MutableList<TargetType>.bindConverted(sourceList: ObservableList<SourceType>, converter: (SourceType) -> TargetType): ListConversionListener<SourceType, TargetType> {
+fun <SourceType, TargetType> MutableList<TargetType>.bind(sourceList: ObservableList<SourceType>, converter: (SourceType) -> TargetType): ListConversionListener<SourceType, TargetType> {
     val listener = ListConversionListener(this, converter)
     if (this is ObservableList<*>) {
         (this as ObservableList<TargetType>).setAll(sourceList.map(converter))